requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==1.11
django-el-pagination==3.2.4
django-material==1.3.0
djangorestframework==3.8.2
idna==2.7
Markdown==3.0.1
MySQL-python==1.2.5
Pillow==5.2.0
python-magic==0.4.15
pytz==2018.5
qrcode==5.1
requests==2.20.0
six==1.11.0
urllib3==1.24
xlrd==1.1.0
xlwt==1.3.0
